refactor: split create course form into components to reduce size
