+++ app/(protected)/courses/new/_components/CourseStructureBuilder.tsx
@@ -1,353 +1,260 @@
 "use client";
 
-import { useState } from "react";
-import { ChevronDown, ChevronRight, Grip, Plus, Trash2 } from "lucide-react";
-import { Button } from "@/components/ui/button";
-import { Card, CardHeader, CardContent } from "@/components/ui/card";
-import { Input } from "@/components/ui/input";
-import { Textarea } from "@/components/ui/textarea";
-import {
-  Accordion,
-  AccordionItem,
-  AccordionTrigger,
-  AccordionContent,
-} from "@/components/ui/accordion";
-import { Volume, Module, Lesson, CourseStructure } from "@/lib/types/course";
+import React, { useState } from "react";
+import { cn } from "@/lib/utils";
+import type { TreeNode, NodeType } from "@/lib/types/course-structure";
 
-interface CourseStructureBuilderProps {
-  value: CourseStructure;
-  onChange: (value: CourseStructure) => void;
+import { Badge } from "@/components/ui/badge";
+import { Button } from "@/components/ui/button";
+import { Input } from "@/components/ui/input";
+import {
+  DropdownMenu,
+  DropdownMenuTrigger,
+  DropdownMenuContent,
+  DropdownMenuItem,
+} from "@/components/ui/dropdown-menu";
+import {
+  Folder,
+  BookOpen,
+  FileText,
+  Trash2,
+  Edit2,
+  Check,
+  X,
+  ChevronDown,
+  ChevronRight,
+  Plus,
+} from "lucide-react";
+
+export function allowedChildTypes(type: NodeType | null): NodeType[] {
+  if (type === null) return ["module", "lesson"];
+  if (type === "module") return ["chapter", "lesson"];
+  if (type === "chapter") return ["lesson"];
+  return [];
 }
 
-export function CourseStructureBuilder({
-  value = { volumes: [] },
+export function EnhancedTreeBuilder({
+  nodes,
   onChange,
-}: CourseStructureBuilderProps) {
-  const [expandedVolume, setExpandedVolume] = useState<string | null>(null);
-  const [expandedModule, setExpandedModule] = useState<string | null>(null);
-
-  const generateId = () => Math.random().toString(36).substring(2, 9);
-
-  const addVolume = () => {
-    const newVolume: Volume = {
-      id: generateId(),
-      title: "New Volume",
-      modules: [],
-    };
-    const updated = {
-      ...value,
-      volumes: [...(value.volumes || []), newVolume],
-    };
-    onChange(updated);
-    setExpandedVolume(newVolume.id);
+}: {
+  nodes: TreeNode[];
+  onChange: (nodes: TreeNode[]) => void;
+}) {
+  const updateNode = (id: string, patch: Partial<TreeNode>) => {
+    onChange(nodes.map((n) => (n.id === id ? { ...n, ...patch } : n)));
   };
-
-  const addModule = (volumeId: string) => {
-    const newModule: Module = {
-      id: generateId(),
-      title: "New Module",
-      lessons: [],
-    };
-    const updated = {
-      ...value,
-      volumes: value.volumes.map((vol) =>
-        vol.id === volumeId
-          ? { ...vol, modules: [...(vol.modules || []), newModule] }
-          : vol
-      ),
-    };
-    onChange(updated);
-    setExpandedModule(newModule.id);
-  };
-
-  const addLesson = (volumeId: string, moduleId: string) => {
-    const newLesson: Lesson = { id: generateId(), title: "New Lesson" };
-    const updated = {
-      ...value,
-      volumes: value.volumes.map((vol) =>
-        vol.id === volumeId
-          ? {
-              ...vol,
-              modules: vol.modules.map((mod) =>
-                mod.id === moduleId
-                  ? { ...mod, lessons: [...(mod.lessons || []), newLesson] }
-                  : mod
-              ),
-            }
-          : vol
-      ),
-    };
-    onChange(updated);
-  };
-
-  const updateVolumeTitle = (volumeId: string, title: string) => {
-    const updated = {
-      ...value,
-      volumes: value.volumes.map((vol) =>
-        vol.id === volumeId ? { ...vol, title } : vol
-      ),
-    };
-    onChange(updated);
-  };
-
-  const updateModuleTitle = (
-    volumeId: string,
-    moduleId: string,
-    title: string
-  ) => {
-    const updated = {
-      ...value,
-      volumes: value.volumes.map((vol) =>
-        vol.id === volumeId
-          ? {
-              ...vol,
-              modules: vol.modules.map((mod) =>
-                mod.id === moduleId ? { ...mod, title } : mod
-              ),
-            }
-          : vol
-      ),
-    };
-    onChange(updated);
-  };
-
-  const updateLessonTitle = (
-    volumeId: string,
-    moduleId: string,
-    lessonId: string,
-    title: string
-  ) => {
-    const updated = {
-      ...value,
-      volumes: value.volumes.map((vol) =>
-        vol.id === volumeId
-          ? {
-              ...vol,
-              modules: vol.modules.map((mod) =>
-                mod.id === moduleId
-                  ? {
-                      ...mod,
-                      lessons: mod.lessons.map((les) =>
-                        les.id === lessonId ? { ...les, title } : les
-                      ),
-                    }
-                  : mod
-              ),
-            }
-          : vol
-      ),
-    };
-    onChange(updated);
-  };
-
-  const removeVolume = (volumeId: string) => {
-    const updated = {
-      ...value,
-      volumes: value.volumes.filter((vol) => vol.id !== volumeId),
-    };
-    onChange(updated);
-  };
-
-  const removeModule = (volumeId: string, moduleId: string) => {
-    const updated = {
-      ...value,
-      volumes: value.volumes.map((vol) =>
-        vol.id === volumeId
-          ? {
-              ...vol,
-              modules: vol.modules.filter((mod) => mod.id !== moduleId),
-            }
-          : vol
-      ),
-    };
-    onChange(updated);
-  };
-
-  const removeLesson = (
-    volumeId: string,
-    moduleId: string,
-    lessonId: string
-  ) => {
-    const updated = {
-      ...value,
-      volumes: value.volumes.map((vol) =>
-        vol.id === volumeId
-          ? {
-              ...vol,
-              modules: vol.modules.map((mod) =>
-                mod.id === moduleId
-                  ? {
-                      ...mod,
-                      lessons: mod.lessons.filter((les) => les.id !== lessonId),
-                    }
-                  : mod
-              ),
-            }
-          : vol
-      ),
-    };
-    onChange(updated);
+  const removeNode = (id: string) => {
+    onChange(nodes.filter((n) => n.id !== id));
   };
 
   return (
-    <div className="space-y-4">
-      <Accordion
-        type="single"
-        collapsible
-        className="w-full"
-        value={expandedVolume || undefined}
-        onValueChange={(val) => setExpandedVolume(val || null)}
-      >
-        {value.volumes.map((volume) => (
-          <AccordionItem key={volume.id} value={volume.id}>
-            <AccordionTrigger className="hover:bg-muted/50 px-4 rounded-md">
-              <div className="flex items-center gap-2">
-                <Grip className="h-4 w-4 text-muted-foreground" />
-                <span>{volume.title || "Untitled Volume"}</span>
-                <span className="text-xs text-muted-foreground">
-                  ({volume.modules.length} modules)
+    <div className="space-y-2">
+      {nodes.map((node) => (
+        <TreeNodeItem
+          key={node.id}
+          node={node}
+          updateNode={updateNode}
+          removeNode={removeNode}
+        />
+      ))}
+    </div>
+  );
+}
+
+function TreeNodeItem({
+  node,
+  updateNode,
+  removeNode,
+}: {
+  node: TreeNode;
+  updateNode: (id: string, patch: Partial<TreeNode>) => void;
+  removeNode: (id: string) => void;
+}) {
+  const [expanded, setExpanded] = useState(true);
+  const [editing, setEditing] = useState(node.title === "");
+  const [title, setTitle] = useState(node.title);
+
+  const handleSave = () => {
+    updateNode(node.id, { title });
+    setEditing(false);
+  };
+
+  const handleCancel = () => {
+    setTitle(node.title);
+    setEditing(false);
+  };
+
+  const getNodeIcon = () => {
+    switch (node.type) {
+      case "module":
+        return <Folder className="h-4 w-4 text-blue-500" />;
+      case "chapter":
+        return <BookOpen className="h-4 w-4 text-emerald-500" />;
+      case "lesson":
+        return <FileText className="h-4 w-4 text-amber-500" />;
+      default:
+        return null;
+    }
+  };
+
+  const getNodeColor = () => {
+    switch (node.type) {
+      case "module":
+        return "border-blue-200 bg-blue-50 dark:border-blue-900 dark:bg-blue-950/30";
+      case "chapter":
+        return "border-emerald-200 bg-emerald-50 dark:border-emerald-900 dark:bg-emerald-950/30";
+      case "lesson":
+        return "border-amber-200 bg-amber-50 dark:border-amber-900 dark:bg-amber-950/30";
+      default:
+        return "border-gray-200 bg-gray-50 dark:border-gray-800 dark:bg-gray-900/30";
+    }
+  };
+
+  return (
+    <div className={cn("border rounded-md overflow-hidden", getNodeColor())}>
+      <div className="flex items-center p-2 gap-2">
+        <button
+          type="button"
+          onClick={() => setExpanded(!expanded)}
+          className="p-1 rounded-sm hover:bg-black/5 dark:hover:bg-white/5"
+          disabled={node.children.length === 0}
+        >
+          {node.children.length > 0 ? (
+            expanded ? (
+              <ChevronDown className="h-4 w-4 text-muted-foreground" />
+            ) : (
+              <ChevronRight className="h-4 w-4 text-muted-foreground" />
+            )
+          ) : (
+            <div className="w-4" />
+          )}
+        </button>
+        <div className="flex items-center gap-1.5">
+          {getNodeIcon()}
+          <Badge
+            variant="outline"
+            className="text-xs font-normal capitalize px-1.5 py-0"
+          >
+            {node.type}
+          </Badge>
+        </div>
+        {editing ? (
+          <div className="flex-1 flex items-center gap-1">
+            <Input
+              value={title}
+              onChange={(e) => setTitle(e.target.value)}
+              placeholder={`${node.type} title`}
+              className="h-8 flex-1"
+              autoFocus
+            />
+            <Button
+              type="button"
+              size="icon"
+              variant="ghost"
+              onClick={handleSave}
+              className="h-8 w-8"
+            >
+              <Check className="h-4 w-4" />
+            </Button>
+            <Button
+              type="button"
+              size="icon"
+              variant="ghost"
+              onClick={handleCancel}
+              className="h-8 w-8"
+            >
+              <X className="h-4 w-4" />
+            </Button>
+          </div>
+        ) : (
+          <>
+            <span className="flex-1 font-medium truncate">
+              {node.title || (
+                <span className="italic text-muted-foreground">
+                  Untitled {node.type}
                 </span>
-              </div>
-            </AccordionTrigger>
-            <AccordionContent className="px-4 pt-2">
-              <div className="space-y-4">
-                <Input
-                  placeholder="Volume Title"
-                  value={volume.title}
-                  onChange={(e) => updateVolumeTitle(volume.id, e.target.value)}
-                />
-                <Textarea
-                  placeholder="Volume Description (optional)"
-                  value={volume.description || ""}
-                  onChange={(e) => updateVolumeTitle(volume.id, e.target.value)}
-                  className="resize-y"
-                />
-                <div className="pl-6 space-y-2">
-                  <div className="font-medium text-sm">Modules</div>
-                  {volume.modules.map((module) => (
-                    <Card key={module.id} className="border border-muted">
-                      <CardHeader className="p-3 flex justify-between items-center">
-                        <div
-                          className="flex items-center gap-2 cursor-pointer"
+              )}
+            </span>
+            <div className="flex items-center gap-1">
+              <Button
+                type="button"
+                size="icon"
+                variant="ghost"
+                onClick={() => setEditing(true)}
+                className="h-8 w-8"
+              >
+                <Edit2 className="h-4 w-4" />
+              </Button>
+              {allowedChildTypes(node.type).length > 0 && (
+                <DropdownMenu>
+                  <DropdownMenuTrigger asChild>
+                    <Button
+                      type="button"
+                      size="icon"
+                      variant="ghost"
+                      className="h-8 w-8"
+                    >
+                      <Plus className="h-4 w-4" />
+                    </Button>
+                  </DropdownMenuTrigger>
+                  <DropdownMenuContent align="end">
+                    {allowedChildTypes(node.type).map((type) => (
+                      <DropdownMenuItem asChild key={type}>
+                        <Button
+                          type="button"
+                          variant="ghost"
+                          className="w-full text-left"
                           onClick={() =>
-                            setExpandedModule(
-                              expandedModule === module.id ? null : module.id
-                            )
+                            updateNode(node.id, {
+                              children: [
+                                ...node.children,
+                                {
+                                  id: crypto.randomUUID(),
+                                  type,
+                                  title: "",
+                                  children: [],
+                                },
+                              ],
+                            })
                           }
                         >
-                          {expandedModule === module.id ? (
-                            <ChevronDown className="h-4 w-4" />
-                          ) : (
-                            <ChevronRight className="h-4 w-4" />
-                          )}
-                          <span>{module.title || "Untitled Module"}</span>
-                          <span className="text-xs text-muted-foreground">
-                            ({module.lessons.length} lessons)
-                          </span>
-                        </div>
-                        <Button
-                          variant="ghost"
-                          size="icon"
-                          onClick={() => removeModule(volume.id, module.id)}
-                        >
-                          <Trash2 className="h-4 w-4" />
-                        </Button>
-                      </CardHeader>
-                      {expandedModule === module.id && (
-                        <CardContent className="p-3 pt-0 space-y-4">
-                          <Input
-                            placeholder="Module Title"
-                            value={module.title}
-                            onChange={(e) =>
-                              updateModuleTitle(
-                                volume.id,
-                                module.id,
-                                e.target.value
-                              )
-                            }
-                          />
-                          <div className="space-y-2">
-                            {module.lessons.map((lesson) => (
-                              <div
-                                key={lesson.id}
-                                className="flex items-center justify-between p-2 border rounded-md"
-                              >
-                                <div className="flex items-center gap-2">
-                                  <Grip className="h-4 w-4 text-muted-foreground" />
-                                  <Input
-                                    value={lesson.title}
-                                    onChange={(e) =>
-                                      updateLessonTitle(
-                                        volume.id,
-                                        module.id,
-                                        lesson.id,
-                                        e.target.value
-                                      )
-                                    }
-                                    className="border-0 p-0 shadow-none focus-visible:ring-0"
-                                  />
-                                </div>
-                                <Button
-                                  variant="ghost"
-                                  size="icon"
-                                  onClick={() =>
-                                    removeLesson(
-                                      volume.id,
-                                      module.id,
-                                      lesson.id
-                                    )
-                                  }
-                                >
-                                  <Trash2 className="h-4 w-4" />
-                                </Button>
-                              </div>
-                            ))}
-                            <Button
-                              variant="outline"
-                              size="sm"
-                              className="w-full"
-                              onClick={() => addLesson(volume.id, module.id)}
-                            >
-                              <Plus className="h-4 w-4 mr-2" /> Add Lesson
-                            </Button>
+                          <div className="flex items-center gap-2">
+                            {type === "module" && (
+                              <Folder className="h-4 w-4" />
+                            )}
+                            {type === "chapter" && (
+                              <BookOpen className="h-4 w-4" />
+                            )}
+                            {type === "lesson" && (
+                              <FileText className="h-4 w-4" />
+                            )}
+                            <span className="capitalize">Add {type}</span>
                           </div>
-                        </CardContent>
-                      )}
-                    </Card>
-                  ))}
-                  <Button
-                    variant="outline"
-                    size="sm"
-                    className="w-full"
-                    onClick={() => addModule(volume.id)}
-                  >
-                    <Plus className="h-4 w-4 mr-2" /> Add Module
-                  </Button>
-                </div>
-                <div className="flex justify-end">
-                  <Button
-                    variant="destructive"
-                    size="sm"
-                    onClick={() => removeVolume(volume.id)}
-                  >
-                    <Trash2 className="h-4 w-4 mr-2" /> Remove Volume
-                  </Button>
-                </div>
-              </div>
-            </AccordionContent>
-          </AccordionItem>
-        ))}
-      </Accordion>
-      <Button variant="outline" className="w-full" onClick={addVolume}>
-        <Plus className="h-4 w-4 mr-2" /> Add Volume
-      </Button>
-      {value.volumes.length === 0 && (
-        <div className="text-center py-8 text-muted-foreground">
-          <div className="mb-2">No volumes added yet</div>
-          <div className="text-sm">
-            Start by adding a volume, then add modules and lessons to create
-            your course structure.
-          </div>
+                        </Button>
+                      </DropdownMenuItem>
+                    ))}
+                  </DropdownMenuContent>
+                </DropdownMenu>
+              )}
+              <Button
+                type="button"
+                size="icon"
+                variant="ghost"
+                onClick={() => removeNode(node.id)}
+                className="h-8 w-8 text-destructive hover:text-destructive"
+              >
+                <Trash2 className="h-4 w-4" />
+              </Button>
+            </div>
+          </>
+        )}
+      </div>
+      {expanded && node.children.length > 0 && (
+        <div className="pl-8 pr-2 pb-2">
+          <EnhancedTreeBuilder
+            nodes={node.children}
+            onChange={(children) => updateNode(node.id, { children })}
+          />
         </div>
       )}
     </div>
